requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Validation & Serialization
marshmallow==3.20.1

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Logging
python-json-logger==2.0.7
